adding inode interface

+++ kernel/pipe.c
@@ -3,8 +3,8 @@
 
 struct Pipe {
     char data[PIPE_SIZE];
-    uint16_t read_index;
-    uint16_t write_index;  
+    uint8_t  read_index;
+    uint8_t  write_index;  
     uint8_t  read_open; 
     uint8_t  write_open; 
 };
